requirements.txt
selenium
pandas
apache-airflow
jsonschema
pytest
